+++ src/main/java/christmas/controller/ReservationController.java
@@ -4,7 +4,7 @@ import static christmas.view.OutputPrinter.OUTPUT_ORIGIN_PRICE_MESSAGE;
 import static christmas.view.OutputPrinter.OUTPUT_TOTAL_PRICE_MESSAGE;
 import static christmas.view.OutputPrinter.OUTPUT_TOTAL_PROFIT_MESSAGE;
 
-import christmas.domain.FoodItem;
+import christmas.domain.food.FoodItem;
 import christmas.domain.event.Grade;
 import christmas.domain.order.Orders;
 import christmas.domain.price.money.KoreaMoney;
@@ -34,18 +34,18 @@ public class ReservationController {
         Orders orders = orderService.createOrder();
         OutputManager.displayOrder(orders);
 
-        Money originPrice = eventService.getOriginPrice(orders);
+        Money originPrice = priceService.getOriginPrice(orders);
         OutputManager.displayMoney(originPrice, OUTPUT_ORIGIN_PRICE_MESSAGE);
 
         List<FoodItem> gifts = eventService.receiveGift(orders);
         OutputManager.displayGifts(gifts);
 
-        Money giftPrice = priceService.applyGift(gifts);
-        Money discountPrice = priceService.applyDiscount(orders);
+        Money giftPrice = eventService.getGiftProfits(gifts);
+        Money discountPrice = eventService.getDiscountProfits(orders);
         OutputManager.displayMoney(KoreaMoney.add(giftPrice, discountPrice), OUTPUT_TOTAL_PROFIT_MESSAGE);
 
-        Money discountedPrice = priceService.getDiscountedPrice(originPrice, discountPrice);
-        OutputManager.displayMoney(discountedPrice, OUTPUT_TOTAL_PRICE_MESSAGE);
+        Money totaldPrice = priceService.getTotalPrice(originPrice, discountPrice);
+        OutputManager.displayMoney(totaldPrice, OUTPUT_TOTAL_PRICE_MESSAGE);
 
         Grade grade = rateService.determineGrade(KoreaMoney.add(giftPrice, discountPrice));
         OutputManager.displayGrade(grade);
