+++ src/main/java/christmas/domain/event/policy/gift/FoodGiftPolicy.java
@@ -1,6 +1,6 @@
 package christmas.domain.event.policy.gift;
 
-import christmas.domain.FoodItem;
+import christmas.domain.food.FoodItem;
 import christmas.domain.event.EventCondition;
 import christmas.domain.order.Orders;
 import java.util.Optional;
