+++ src/main/java/christmas/domain/food/Quantity.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.food;
 
 import christmas.validator.OrderValidator;
