+++ src/main/java/christmas/domain/price/PriceCalculator.java
@@ -1,6 +1,6 @@
 package christmas.domain.price;
 
-import christmas.domain.FoodItem;
+import christmas.domain.food.FoodItem;
 import christmas.domain.event.policy.PolicyManager;
 import christmas.domain.order.Orders;
 import java.util.Collections;
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+//TODO 역할 분리
 public class PriceCalculator {
     private PolicyManager policyManager;
