+++ src/main/java/christmas/view/OutputManager.java
@@ -5,7 +5,7 @@ import static christmas.view.OutputPrinter.OUTPUT_ORDER_MESSAGE;
 import static christmas.view.OutputPrinter.newLine;
 import static christmas.view.OutputPrinter.printMessage;
 
-import christmas.domain.FoodItem;
+import christmas.domain.food.FoodItem;
 import christmas.domain.event.Grade;
 import christmas.domain.order.Orders;
 import christmas.domain.price.money.Money;
