+++ src/test/java/christmas/domain/discount/SpecialDiscountPolicyTest.java
@@ -1,6 +1,6 @@
 //package christmas.domain.discount;
 //
-//import christmas.domain.FoodItem;
+//import christmas.domain.event.food.FoodItem;
 //import christmas.domain.order.Orders;
 //import christmas.domain.event.policy.discount.DiscountPolicy;
 //import christmas.domain.event.policy.discount.SpecialDiscountPolicy;
