+++ src/test/java/christmas/domain/event/EventConditionTest.java
@@ -3,7 +3,7 @@
 //import static org.assertj.core.api.Assertions.assertThat;
 //import static org.junit.jupiter.api.Assertions.*;
 //
-//import christmas.domain.FoodItem;
+//import christmas.domain.event.food.FoodItem;
 //import christmas.domain.order.Orders;
 //import org.junit.jupiter.api.Test;
 //
