+++ src/test/java/christmas/domain/price/PriceCalculatorTest.java
@@ -1,6 +1,6 @@
 package christmas.domain.price;
 
-import christmas.domain.FoodItem;
+import christmas.domain.food.FoodItem;
 import christmas.domain.event.policy.PolicyManager;
 import christmas.domain.order.Orders;
 import java.time.LocalDate;
